Modifications to fix warnings during compilation time

+++ src/Derived_types.f90
@@ -12,7 +12,7 @@ MODULE Derived_types
   TYPE wl_pattern_elem
      INTEGER :: ix_relative !Relative x-coordinate of a given element stored in the associated k-wind exposure array
      INTEGER :: jy_relative !Relative y-coordinate of a given element stored in the associated k-wind exposure array
-     REAL :: horizontal_dist !Distance between the given k-wind exposure array's element and the grid point whose WL index is being calculated
+     DOUBLE PRECISION :: horizontal_dist !Distance between the given k-wind exposure array's element and the grid point whose WL index is being calculated
   END TYPE wl_pattern_elem
